requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP and embeddings
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0

# Jupyter (optional - for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API calls (when ready)
openai>=1.0.0
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
